+++ homematic-devices/hm-cc-rt-dn.js
@@ -1,4 +1,5 @@
 const Accessory = require('./lib/accessory');
+const thermostat = require('./lib/thermostat');
 
 module.exports = class HmCcRtDn extends Accessory {
     init(config, node) {
@@ -7,7 +8,7 @@ module.exports = class HmCcRtDn extends Accessory {
 
         let valveState = 0;
         let currentSetpoint;
-        let valueSetpoint = 21;
+        const setpoint = thermostat.createSetpoint();
         let controlMode;
         let target;
 
@@ -16,15 +17,15 @@ module.exports = class HmCcRtDn extends Accessory {
             switch (controlMode) {
                 case 1:
                     // Manu
-                    target = currentSetpoint > 4.5 ? 1 : 0;
+                    target = thermostat.targetState(true, currentSetpoint);
                     break;
                 case 0:
                     // Auto
-                    target = 3;
+                    target = thermostat.targetState(false, currentSetpoint);
                     break;
                 case 2:
                     // Party
-                    target = 3;
+                    target = thermostat.targetState(false, currentSetpoint);
                     break;
                 case 3:
                     // Boost
@@ -33,7 +34,7 @@ module.exports = class HmCcRtDn extends Accessory {
                 default:
             }
 
-            return controlMode === 3 ? 1 : target;
+            return controlMode === 3 ? 1 : (target ?? 3);
         }
 
         function currentState() {
@@ -50,14 +51,12 @@ module.exports = class HmCcRtDn extends Accessory {
             .setProps('TargetTemperature', {minValue: 4.5, maxValue: 30.5, minStep: 0.5})
             .get('TargetTemperature', config.deviceAddress + ':4.SET_TEMPERATURE', (value) => {
                 currentSetpoint = value;
-                if (value > 4.5) {
-                    valueSetpoint = value;
-                }
+                setpoint.read(value);
 
                 updateHeatingCoolingState();
                 return value;
             })
-            .set('TargetTemperature', config.deviceAddress + ':4.SET_TEMPERATURE')
+            .set('TargetTemperature', config.deviceAddress + ':4.SET_TEMPERATURE', (value) => setpoint.write(value))
 
             .setProps('CurrentHeatingCoolingState', {validValues: [0, 1], maxValue: 1})
             .get('CurrentHeatingCoolingState', config.deviceAddress + ':4.SET_TEMPERATURE', () => {
@@ -74,35 +73,18 @@ module.exports = class HmCcRtDn extends Accessory {
                 }, 1000);
                 return targetState();
             })
-            .set('TargetHeatingCoolingState', (value, callback) => {
-                // 0=off, 1=heat, 3=auto
-                if (value === 0) {
-                    ccu.setValue(config.iface, config.description.ADDRESS + ':4', 'MANU_MODE', 4.5)
-                        .then(() => {
-                            callback();
-                        })
-                        .catch(() => {
-                            callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                        });
-                } else if (value === 1) {
-                    ccu.setValue(config.iface, config.description.ADDRESS + ':4', 'MANU_MODE', valueSetpoint)
-                        .then(() => {
-                            serviceThermostat.update('TargetTemperature', valueSetpoint);
-                            callback();
-                        })
-                        .catch(() => {
-                            callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                        });
-                } else {
-                    ccu.setValue(config.iface, config.description.ADDRESS + ':4', 'AUTO_MODE', true)
-                        .then(() => {
-                            callback();
-                        })
-                        .catch(() => {
-                            callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                        });
-                }
-            });
+            .set(
+                'TargetHeatingCoolingState',
+                thermostat.hmModeSetter({
+                    ccu,
+                    hap,
+                    node,
+                    config,
+                    channel: 4,
+                    setpoint,
+                    service: serviceThermostat,
+                }),
+            );
 
         function updateHeatingCoolingState() {
             serviceThermostat.update('CurrentHeatingCoolingState', currentState());
@@ -140,7 +122,7 @@ module.exports = class HmCcRtDn extends Accessory {
                                 callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
                             });
                     } else if (target === 0) {
-                        ccu.setValue(config.iface, config.description.ADDRESS + ':4', 'MANU_MODE', valueSetpoint)
+                        ccu.setValue(config.iface, config.description.ADDRESS + ':4', 'MANU_MODE', setpoint.value)
                             .then(() => {
                                 callback();
                             })
@@ -148,7 +130,7 @@ module.exports = class HmCcRtDn extends Accessory {
                                 callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
                             });
                     } else if (target === 1) {
-                        ccu.setValue(config.iface, config.description.ADDRESS + ':4', 'MANU_MODE', valueSetpoint)
+                        ccu.setValue(config.iface, config.description.ADDRESS + ':4', 'MANU_MODE', setpoint.value)
                             .then(() => {
                                 callback();
                             })