+++ homematic-devices/hm-cc-vg-1.js
@@ -1,4 +1,5 @@
 const Accessory = require('./lib/accessory');
+const thermostat = require('./lib/thermostat');
 
 module.exports = class HmCcVg1 extends Accessory {
     init(config, node) {
@@ -6,7 +7,7 @@ module.exports = class HmCcVg1 extends Accessory {
         const {hap} = bridgeConfig;
 
         let currentSetpoint;
-        let valueSetpoint = 21;
+        const setpoint = thermostat.createSetpoint();
 
         let controlMode;
         let target;
@@ -65,15 +66,15 @@ module.exports = class HmCcVg1 extends Accessory {
             switch (controlMode) {
                 case 1:
                     // Manu
-                    target = currentSetpoint > 4.5 ? 1 : 0;
+                    target = thermostat.targetState(true, currentSetpoint);
                     break;
                 case 0:
                     // Auto
-                    target = 3;
+                    target = thermostat.targetState(false, currentSetpoint);
                     break;
                 case 2:
                     // Party
-                    target = 3;
+                    target = thermostat.targetState(false, currentSetpoint);
                     break;
                 case 3:
                     // Boost
@@ -82,7 +83,7 @@ module.exports = class HmCcVg1 extends Accessory {
                 default:
             }
 
-            return controlMode === 3 ? 1 : target;
+            return controlMode === 3 ? 1 : (target ?? 3);
         }
 
         function currentState() {
@@ -100,14 +101,12 @@ module.exports = class HmCcVg1 extends Accessory {
             .setProps('TargetTemperature', {minValue: 4.5, maxValue: 30.5, minStep: 0.5})
             .get('TargetTemperature', valueChannel + '.SET_TEMPERATURE', (value) => {
                 currentSetpoint = value;
-                if (value > 4.5) {
-                    valueSetpoint = value;
-                }
+                setpoint.read(value);
 
                 updateHeatingCoolingState();
                 return value;
             })
-            .set('TargetTemperature', config.deviceAddress + ':1.SET_TEMPERATURE')
+            .set('TargetTemperature', config.deviceAddress + ':1.SET_TEMPERATURE', (value) => setpoint.write(value))
 
             .setProps('CurrentHeatingCoolingState', {validValues: [0, 1], maxValue: 1})
             .get('CurrentHeatingCoolingState', valueChannel + '.SET_TEMPERATURE', () => {
@@ -124,38 +123,21 @@ module.exports = class HmCcVg1 extends Accessory {
                 }, 1000);
                 return targetState();
             })
-            .set('TargetHeatingCoolingState', (value, callback) => {
-                // 0=off, 1=heat, 3=auto
-                if (value === 0) {
-                    ccu.setValue(config.iface, config.description.ADDRESS + ':1', 'MANU_MODE', 4.5)
-                        .then(() => {
-                            controlMode = 1;
-                            callback();
-                        })
-                        .catch(() => {
-                            callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                        });
-                } else if (value === 1) {
-                    ccu.setValue(config.iface, config.description.ADDRESS + ':1', 'MANU_MODE', valueSetpoint)
-                        .then(() => {
-                            serviceThermostat.update('TargetTemperature', valueSetpoint);
-                            controlMode = 1;
-                            callback();
-                        })
-                        .catch(() => {
-                            callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                        });
-                } else {
-                    ccu.setValue(config.iface, config.description.ADDRESS + ':1', 'AUTO_MODE', true)
-                        .then(() => {
-                            controlMode = 0;
-                            callback();
-                        })
-                        .catch(() => {
-                            callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                        });
-                }
-            });
+            .set(
+                'TargetHeatingCoolingState',
+                thermostat.hmModeSetter({
+                    ccu,
+                    hap,
+                    node,
+                    config,
+                    channel: 1,
+                    setpoint,
+                    service: serviceThermostat,
+                    onMode: (mode) => {
+                        controlMode = mode;
+                    },
+                }),
+            );
 
         function updateHeatingCoolingState() {
             const current = currentState();
@@ -233,7 +215,7 @@ module.exports = class HmCcVg1 extends Accessory {
                         dp = 'BOOST_MODE';
                     } else if (target === 0 || target === 1) {
                         dp = 'MANU_MODE';
-                        value = valueSetpoint;
+                        value = setpoint.value;
                     } else {
                         dp = 'AUTO_MODE';
                         value = true;