+++ homematic-devices/hmip-sth.js
@@ -1,4 +1,5 @@
 const Accessory = require('./lib/accessory');
+const thermostat = require('./lib/thermostat');
 
 module.exports = class HmipWth extends Accessory {
     init(config, node) {
@@ -8,7 +9,7 @@ module.exports = class HmipWth extends Accessory {
         const levels = {};
         let level = 0;
         let currentSetpoint;
-        let valueSetpoint = 21;
+        const setpoint = thermostat.createSetpoint();
         let setpointMode;
         let target;
         let serviceThermostat;
@@ -18,11 +19,11 @@ module.exports = class HmipWth extends Accessory {
             switch (setpointMode) {
                 case 1:
                     // Manu
-                    target = currentSetpoint > 4.5 ? 1 : 0;
+                    target = thermostat.targetState(true, currentSetpoint);
                     break;
                 default:
                     // Auto / Party
-                    target = 3;
+                    target = thermostat.targetState(false, currentSetpoint);
             }
 
             return target;
@@ -43,7 +44,6 @@ module.exports = class HmipWth extends Accessory {
             node.debug(config.name + ' linked to ' + JSON.stringify(links));
 
             serviceThermostat = this.addService('Thermostat', config.name);
-            const subtypeThermostat = serviceThermostat.subtype;
 
             serviceThermostat
                 .setProps('CurrentTemperature', {minValue: -40, maxValue: 80})
@@ -52,14 +52,14 @@ module.exports = class HmipWth extends Accessory {
                 .setProps('TargetTemperature', {minValue: 4.5, maxValue: 30.5, minStep: 0.5})
                 .get('TargetTemperature', config.deviceAddress + ':1.SET_POINT_TEMPERATURE', (value) => {
                     currentSetpoint = value;
-                    if (value !== 4.5) {
-                        valueSetpoint = value;
-                    }
+                    setpoint.read(value);
 
                     updateHeatingCoolingState();
                     return value;
                 })
-                .set('TargetTemperature', config.deviceAddress + ':1.SET_POINT_TEMPERATURE')
+                .set('TargetTemperature', config.deviceAddress + ':1.SET_POINT_TEMPERATURE', (value) =>
+                    setpoint.write(value),
+                )
 
                 .setProps('CurrentHeatingCoolingState', {validValues: [0, 1], maxValue: 1})
                 .get('CurrentHeatingCoolingState', config.deviceAddress + ':1.LEVEL', () => {
@@ -76,98 +76,20 @@ module.exports = class HmipWth extends Accessory {
                     }, 1000);
                     return targetState();
                 })
-                .set('TargetHeatingCoolingState', (value, callback) => {
-                    // 0=off, 1=heat, 3=auto
-                    if (value === 0) {
-                        const params = {
-                            CONTROL_MODE: 1,
-                            SET_POINT_TEMPERATURE: 4.5,
-                        };
-                        node.debug(
-                            'set ' +
-                                config.name +
-                                ' (' +
-                                subtypeThermostat +
-                                ') TargetHeatingCoolingState ' +
-                                value +
-                                ' -> ' +
-                                config.description.ADDRESS +
-                                ':1 ' +
-                                JSON.stringify(params),
-                        );
-
-                        ccu.methodCall(config.iface, 'putParamset', [
-                            config.description.ADDRESS + ':1',
-                            'VALUES',
-                            params,
-                        ])
-                            .then(() => {
-                                callback();
-                            })
-                            .catch(() => {
-                                callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                            });
-                    } else if (value === 1) {
-                        if (setpointMode === 1) {
-                            callback();
-                        } else {
-                            const params = {
-                                CONTROL_MODE: 1,
-                                SET_POINT_TEMPERATURE: valueSetpoint,
-                            };
-
-                            node.debug(
-                                'set ' +
-                                    config.name +
-                                    ' (' +
-                                    subtypeThermostat +
-                                    ') TargetHeatingCoolingState ' +
-                                    value +
-                                    ' -> ' +
-                                    config.description.ADDRESS +
-                                    ':1 ' +
-                                    JSON.stringify(params),
-                            );
-                            ccu.methodCall(config.iface, 'putParamset', [
-                                config.description.ADDRESS + ':1',
-                                'VALUES',
-                                params,
-                            ])
-                                .then(() => {
-                                    serviceThermostat.update('TargetTemperature', valueSetpoint);
-                                    callback();
-                                })
-                                .catch(() => {
-                                    callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                                });
-                        }
-                    } else {
-                        const val = value === 3 ? 0 : 1;
-                        if (setpointMode === val) {
-                            callback();
-                        } else {
-                            node.debug(
-                                'set ' +
-                                    config.name +
-                                    ' (' +
-                                    subtypeThermostat +
-                                    ') TargetHeatingCoolingState ' +
-                                    value +
-                                    ' -> ' +
-                                    config.description.ADDRESS +
-                                    ':1.CONTROL_MODE ' +
-                                    (value === 3 ? 0 : 1),
-                            );
-                            ccu.setValue(config.iface, config.description.ADDRESS + ':1', 'CONTROL_MODE', val)
-                                .then(() => {
-                                    callback();
-                                })
-                                .catch(() => {
-                                    callback(new hap.HapStatusError(hap.HAPStatus.SERVICE_COMMUNICATION_FAILURE));
-                                });
-                        }
-                    }
-                });
+                .set(
+                    'TargetHeatingCoolingState',
+                    thermostat.hmipModeSetter({
+                        ccu,
+                        hap,
+                        node,
+                        config,
+                        channel: 1,
+                        setpoint,
+                        service: serviceThermostat,
+                        getMode: () => setpointMode,
+                        getSetpoint: () => currentSetpoint,
+                    }),
+                );
 
             links.forEach((link) => {
                 const valveStateDevice = link.split(':')[0];